analytics/src/analytics.py
"""
Sleepr Analytics Module
Main module for fantasy football analytics and machine learning predictions
"""

import os
import logging
from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

# Database configuration
DATABASE_URL = os.getenv('DATABASE_URL', 'postgresql://localhost/sleepr')
engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

def get_db():
    """Get database session"""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

class SleeprAnalytics:
    """Main analytics class for Sleepr fantasy football analysis"""
    
    def __init__(self):
        self.engine = engine
        self.logger = logger
        
    def analyze_player_performance(self, player_id: int, weeks_back: int = 8):
        """
        Analyze player performance trends
        
        Args:
            player_id: ID of the player to analyze
            weeks_back: Number of weeks to look back for analysis
            
        Returns:
            dict: Analysis results including trends, projections, and recommendations
        """
        # Placeholder for ML analysis
        return {
            'player_id': player_id,
            'trend': 'upward',
            'projected_points': 15.2,
            'confidence': 0.78,
            'recommendation': 'hold'
        }
    
    def get_waiver_wire_targets(self, league_id: str, position: str = None):
        """
        Identify high-value waiver wire targets
        
        Args:
            league_id: League ID to analyze
            position: Optional position filter
            
        Returns:
            list: List of recommended waiver wire pickups
        """
        # Placeholder for waiver wire analysis
        return [
            {
                'player_id': 123,
                'name': 'Breakout Candidate',
                'position': 'RB',
                'opportunity_score': 8.5,
                'reason': 'Increasing snap count and red zone touches'
            }
        ]
    
    def analyze_trade_value(self, player_ids: list):
        """
        Analyze trade values for multiple players
        
        Args:
            player_ids: List of player IDs to evaluate
            
        Returns:
            dict: Trade values and recommendations
        """
        # Placeholder for trade analysis
        return {
            'trade_values': {pid: 8.5 for pid in player_ids},
            'recommendation': 'fair_trade',
            'confidence': 0.82
        }
    
    def optimize_lineup(self, team_id: int, week: int):
        """
        Optimize lineup for maximum projected points
        
        Args:
            team_id: Team ID to optimize
            week: Week number to optimize for
            
        Returns:
            dict: Optimized lineup with projections
        """
        # Placeholder for lineup optimization
        return {
            'team_id': team_id,
            'week': week,
            'optimized_lineup': {
                'QB': {'player_id': 1, 'projected_points': 20.5},
                'RB1': {'player_id': 2, 'projected_points': 18.2},
                'RB2': {'player_id': 3, 'projected_points': 12.8},
                'WR1': {'player_id': 4, 'projected_points': 16.4},
                'WR2': {'player_id': 5, 'projected_points': 14.1},
                'TE': {'player_id': 6, 'projected_points': 10.3},
                'FLEX': {'player_id': 7, 'projected_points': 11.7},
                'DST': {'player_id': 8, 'projected_points': 8.5},
                'K': {'player_id': 9, 'projected_points': 7.2}
            },
            'total_projected': 119.7
        }
    
    def dynasty_asset_analysis(self, team_id: int):
        """
        Analyze dynasty team assets and provide recommendations
        
        Args:
            team_id: Dynasty team ID to analyze
            
        Returns:
            dict: Dynasty analysis and recommendations
        """
        # Placeholder for dynasty analysis
        return {
            'team_id': team_id,
            'young_assets': ['Player A', 'Player B'],
            'aging_assets': ['Player C', 'Player D'],
            'trade_recommendations': [
                {
                    'action': 'sell',
                    'player': 'Aging RB',
                    'reason': 'Declining usage and age concerns',
                    'urgency': 'high'
                }
            ],
            'draft_strategy': 'focus_on_youth'
        }

if __name__ == "__main__":
    analytics = SleeprAnalytics()
    print("Sleepr Analytics initialized successfully")
