analytics/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
